fix must select recipes to delete them

+++ src/userInterface/EliminationRecipePanel.java
@@ -49,31 +49,35 @@ public class EliminationRecipePanel extends JPanel implements ActionListener {
     public void actionPerformed(ActionEvent event) {
         String source = event.getActionCommand();
         if (source.equals("Delete")) {
-            try {
-                int confirmation = JOptionPane.showConfirmDialog(
-                        null,
-                        "Do you really want to delete this recipe(s) ?",
-                        "This is permanent !",
-                        JOptionPane.YES_NO_OPTION);
-                if (confirmation == OK_OPTION) {
-                    ArrayList<Recipe> recipes = recipeController.getAllRecipe();
-                    int[] selectedIndex = this.listSelectionModelRecipe.getSelectedIndices();
-                    for (int index = 0; index < selectedIndex.length; index++) {
-                        int codeRecipe = recipes.get(index).getCode();
-                        this.recipeStepController.deleteAllStepRecipe(codeRecipe);
-                        this.lineRecipeController.deleteAllLineRecip(codeRecipe);
-                        this.equipementController.deleteAllEquipement(codeRecipe);
-                        this.commentController.deleteAllComment(codeRecipe);
-                        this.periodController.deleteAllPeriod(codeRecipe);
-                        this.orderTypeController.deleteAllOrder(codeRecipe);
-                        this.recipeController.deleteRecipe(codeRecipe);
+            int[] selectedIndex = this.listSelectionModelRecipe.getSelectedIndices();
+            if (selectedIndex.length > 0) {
+                try {
+                    int confirmation = JOptionPane.showConfirmDialog(
+                            null,
+                            "Do you really want to delete this recipe(s) ?",
+                            "This is permanent !",
+                            JOptionPane.YES_NO_OPTION);
+                    if (confirmation == OK_OPTION) {
+                        ArrayList<Recipe> recipes = recipeController.getAllRecipe();
+                        for (int index = 0; index < selectedIndex.length; index++) {
+                            int codeRecipe = recipes.get(index).getCode();
+                            this.recipeStepController.deleteAllStepRecipe(codeRecipe);
+                            this.lineRecipeController.deleteAllLineRecip(codeRecipe);
+                            this.equipementController.deleteAllEquipement(codeRecipe);
+                            this.commentController.deleteAllComment(codeRecipe);
+                            this.periodController.deleteAllPeriod(codeRecipe);
+                            this.orderTypeController.deleteAllOrder(codeRecipe);
+                            this.recipeController.deleteRecipe(codeRecipe);
+                        }
+                        this.listSelectionModelRecipe.clearSelection();
+                        this.repaint();
+                        this.revalidate();
                     }
-                    this.listSelectionModelRecipe.clearSelection();
-                    this.repaint();
-                    this.revalidate();
+                } catch (Exception exception) {
+                    JOptionPane.showMessageDialog(null, "Failed to delete recipe", "Failed To Delete recipe", JOptionPane.ERROR_MESSAGE);
                 }
-            } catch (Exception exception) {
-                JOptionPane.showMessageDialog(null, "Failed to delete recipe", "Failed To Delete recipe", JOptionPane.ERROR_MESSAGE);
+            } else {
+                JOptionPane.showMessageDialog(null, "You must select recipes", "Select Recipes", JOptionPane.ERROR_MESSAGE);
             }
         }
     }
